+++ packages/inquiry/template-builders/index.js
@@ -1,0 +1,7 @@
+const confirmationBuilder = require('./confirmation-builder');
+const notificationBuilder = require('./notification-builder');
+
+module.exports = {
+  confirmationBuilder,
+  notificationBuilder,
+};
